realtime_transcription/requirements.txt
deepgram-sdk>=2.11.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
asyncio>=3.4.3